+++ pocket_ic/__init__.py
@@ -1,11 +1,11 @@
 """
-'PocketIC' represents an IC instance and is the public interface
+`PocketIC` represents an IC instance and is the public interface
 of this package.
 
 Instances are running on a PocketIC server, which is represented
-by 'PocketICServer'.
+by `PocketICServer`.
 
-'SubnetConfig' is used to configure the subnets of a PocketIC instance.
+`SubnetConfig` is used to configure the subnets of a PocketIC instance.
 """
 
 from .pocket_ic import *